#3 Added scheduled service for taking webcam snapshots

+++ piwind/src/main/java/com/piotrmajcher/piwind/repositories/ExternalTemperatureRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.piotrmajcher.piwind.domain.ExternalTemperature;
 
-public interface ExternalTemperatureRepository extends CrudRepository<ExternalTemperature, Long>{
+public interface ExternalTemperatureRepository extends CrudRepository<ExternalTemperature, Integer>{
 
 }
