+++ piwind/src/main/java/com/piotrmajcher/piwind/repositories/InternalTemperatureRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.piotrmajcher.piwind.domain.InternalTemperature;
 
-public interface InternalTemperatureRepository extends CrudRepository<InternalTemperature, Long> {
+public interface InternalTemperatureRepository extends CrudRepository<InternalTemperature, Integer> {
 	
 }
